+++ src/components/MyInput/my-input.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import classes from './MyInput.module.css'
+import classes from './my-input.module.css'
 
 export function MyInput(props:any) {
     return <input className={classes.input} {...props}></input>
